retail_image_submission_api/requirements.txt
Flask==0.12.2
Flask-Login==0.4.0
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.2
Jinja2==2.9.6
MarkupSafe==1.0
Pillow==4.1.1
SQLAlchemy==1.1.10
Werkzeug==0.12.2
amqp==1.4.9
aniso8601==1.2.1
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.23
celery==3.1.25
certifi==2018.1.18
click==6.7
cloudinary==1.11.0
et-xmlfile==1.0.1
funcsigs==1.0.2
itsdangerous==0.24
jdcal==1.3
kombu==3.0.37
mock==2.0.0
olefile==0.44
openpyxl==2.5.1
pbr==3.1.1
psycopg2==2.7.1
python-dateutil==2.6.0
pytz==2017.2
six==1.10.0
urllib3==1.22
uWSGI==2.0.15
wsgiref==0.1.2
xlrd==1.1.0
